Move test to core package
